# objfile: optimize findAllOccurrences